Moved code to util
